completando os atributos e metodos

+++ Projeto_OO/src/Republica/Pessoas.java
@@ -6,7 +6,7 @@ public class Pessoas {
 		   genero;
 	
 	int idade;
-	int rendimento;
+	float rendimento;
 	
 	void pessoas() {
